fix: use BigInteger for shares_outstanding to handle IDX share counts
Indonesian stocks like ADRO (31.9B), ENRG (47.1B), and ANTM (24B)
exceed int32 max (~2.1B). Switch to BigInteger (int64) in both the
ORM model and Alembic migration.

+++ data_platform/alembic_migrations/versions/017_commodity_company_profiles.py
@@ -23,7 +23,7 @@ def upgrade() -> None:
         sa.Column("ticker", sa.String(10), nullable=False),
         sa.Column("commodity_type", sa.String(20), nullable=False),
         sa.Column("profile_data", JSONB, nullable=False, server_default=sa.text("'{}'::jsonb")),
-        sa.Column("shares_outstanding", sa.Integer, nullable=False),
+        sa.Column("shares_outstanding", sa.BigInteger, nullable=False),
         sa.Column("trailing_revenue_idr", sa.Float, nullable=False),
         sa.Column("net_margin", sa.Float, nullable=False),
         sa.Column("updated_at", sa.TIMESTAMP(timezone=True), server_default=sa.text("now()")),
@@ -38,7 +38,7 @@ def upgrade() -> None:
         sa.column("ticker", sa.String),
         sa.column("commodity_type", sa.String),
         sa.column("profile_data", JSONB),
-        sa.column("shares_outstanding", sa.Integer),
+        sa.column("shares_outstanding", sa.BigInteger),
         sa.column("trailing_revenue_idr", sa.Float),
         sa.column("net_margin", sa.Float),
     )
